QcomModulePkg: Fix issue related to high flash timing
Optimize fastboot image flash path by storing the boot type instead
of fetching it for every chunk, as is being done currently, since
boot type is not expected to change in current boot. This further helps
to improve time to flash as it reduces an extra uefi call while
flashing each chunk.

Change-Id: I670145991af9b4b50fd9dbafa699eb36044f474a